moved user-info to model folder
